Update Seurat.R
Change 1. Calling 10x function: Since the path has been set up previously it is not needed to paste in the working directory to set the path of our data. Also, I personally find more informative to make clear which are the parameters for newcomers in case it is needed to be modified.

Change 2. Addition of id conversion: In case it is needed, we can use a function to map the id back to the symbol. It might be useful depending on the original data we are working with. Just in case the most possible case, converting ENSEMBL to genesymbol, was defined.

Change 3. Include some annotation in seurat object generation, as well as adding the default parameters definition for clearer understanding of the process.

Change 4. Control check: Just for contrast, added feature according ribosome pattern as well as a adding the sample differenciation earlier in the code. This way, we will be able to visualize the violin plot according to sample. Also add another plot visualization that allows to explore correlation.

Change 5. Identification of highly variable features: Adding a variable feature plot labelled by those 10 genes that are determined to be the most differentially expressed.

Change 6. Clustering visualization: Added a group by sampletype sine it felt like it was missing, as well as visualization for condition and sampletype for both UMAP and tSNE.

Suggestion: Maybe we can create a markdown of this process for a cleaner presentation, which would also allow to show the visual outputs.