Update Seurat.R
Suggestion --> If the sake of this document is to help to understand the pipeline, it could be used the list of top10 generated earlier in the steps regarding the features variable, which would ease the following steps and computational memory needed. Also, it would still be useful to specify that an independent list of genes can be created depending on the basis of the study.